Update
finaliized code to match the companies between the two databases using the zip code, first and second name of the company.  Now I need to add the code that will create the additional columns from the DAT_CIQ dataframe.  Then the code to write to those new cells in the df.