get_match_v4
Recreated yet again the merge formula.  I figured out that I only needed to maintain one loop as the list comprehesion for the SF dataframe could be simplified by specifying  x[#] for x in SF could pull directly the matching variable.  This eliminated the necessity to loop over the SF dataframe and I only needed the 'in' statement to find a match.  Still struggling to get the merge function to work properly.